rename iocage.lib.Config.Resource.ResourceConfig to  iocage.lib.Config.Resource
