+++ iocage/lib/Config/Resource.py
